+++ test/td_cache/templates/acl_loader_test.exs
@@ -7,7 +7,6 @@ defmodule TdCache.Templates.AclLoaderTest do
   alias TdCache.Templates.AclLoader
 
   describe "get_roles_and_users/1" do
-
     setup do
       on_exit(fn -> Redix.del!(["*"]) end)
       :ok
@@ -30,20 +29,19 @@ defmodule TdCache.Templates.AclLoaderTest do
       AclCache.set_acl_role_users("domain", domain2.id, "role3", [user_id_3])
 
       assert %{
-        "role1" => [%{id: ^user_id_1}],
-        "role2" => [%{id: ^user_id_2}],
-      } = AclLoader.get_roles_and_users([domain1.id])
+               "role1" => [%{id: ^user_id_1}],
+               "role2" => [%{id: ^user_id_2}]
+             } = AclLoader.get_roles_and_users([domain1.id])
 
       assert %{
-        "role1" => [%{id: ^user_id_1}],
-        "role2" => [%{id: ^user_id_1}, %{id: ^user_id_2}],
-        "role3" => [%{id: ^user_id_3}],
-      } = AclLoader.get_roles_and_users([domain1.id, domain2.id])
+               "role1" => [%{id: ^user_id_1}],
+               "role2" => [%{id: ^user_id_1}, %{id: ^user_id_2}],
+               "role3" => [%{id: ^user_id_3}]
+             } = AclLoader.get_roles_and_users([domain1.id, domain2.id])
     end
   end
 
   describe "get_roles_and_groups/1" do
-
     setup do
       on_exit(fn -> Redix.del!(["*"]) end)
       :ok
@@ -68,16 +66,15 @@ defmodule TdCache.Templates.AclLoaderTest do
       AclCache.get_acl_role_groups("domain", domain1.id, "role1")
 
       assert %{
-        "role1" => [%{id: ^group_id_1}],
-        "role2" => [%{id: ^group_id_2}],
-      } = AclLoader.get_roles_and_groups([domain1.id])
+               "role1" => [%{id: ^group_id_1}],
+               "role2" => [%{id: ^group_id_2}]
+             } = AclLoader.get_roles_and_groups([domain1.id])
 
       assert %{
-        "role1" => [%{id: ^group_id_1}],
-        "role2" => [%{id: ^group_id_1}, %{id: ^group_id_2}],
-        "role3" => [%{id: ^group_id_3}],
-      } = AclLoader.get_roles_and_groups([domain1.id, domain2.id])
+               "role1" => [%{id: ^group_id_1}],
+               "role2" => [%{id: ^group_id_1}, %{id: ^group_id_2}],
+               "role3" => [%{id: ^group_id_3}]
+             } = AclLoader.get_roles_and_groups([domain1.id, domain2.id])
     end
   end
-
 end